Extract ExecutionPool from LocalPlatform
- Provided default HTTP API for exposing ExecutionPool state to UI
